feat: discord oauth & registration with oauth

+++ Dtos/Responses/RecordResponseDto.cs
@@ -12,7 +12,7 @@ public class RecordResponseDto
 
     public UserDto Player { get; set; } = null!;
 
-    public int ExperienceDelta { get; set; }
+    public ulong ExperienceDelta { get; set; }
 
     public double RksBefore { get; set; }
